requirements.txt
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
openai==1.7.1
pydantic==2.5.3
pydantic_core==2.14.6
pyTelegramBotAPI==4.14.1
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
wheel==0.42.0
